Code cleaning : removed STL strings and string streams
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@932 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/engine/Settings.h
@@ -24,7 +24,7 @@ class Settings
   // Options
   bool QuickWaveRender;
   bool dbWaveRender;
-  unsigned int	maxUndoRedoDepth;
+  unsigned long	maxUndoRedoDepth;
 
   long OutputSystem;
   long InputSystem;
